--- src/main/java/org/example/sqi_images/employee/dto/response/ImageDataResponse.java
@@ -1,7 +1,0 @@
-package org.example.sqi_images.employee.dto.response;
-
-public record ImageDataResponse(byte[] imageData, String contentType) {
-    public static ImageDataResponse of(byte[] imageData, String contentType) {
-        return new ImageDataResponse(imageData, contentType);
-    }
-}
