+++ src/main/java/org/example/sqi_images/photo/dto/response/PhotoDataResponse.java
@@ -1,0 +1,7 @@
+package org.example.sqi_images.photo.dto.response;
+
+public record PhotoDataResponse(byte[] photoData) {
+    public static PhotoDataResponse from(byte[] photoData) {
+        return new PhotoDataResponse(photoData);
+    }
+}
